splitting files for integrating e2e tests.
