+++ lib/FileProcessor.js
